Make serveExternals a boolean parameter on render and renderPageContributions #401
