+++ src/main/resources/lib/enonic/react4xp/pageContributions/appendCssToHeadEnd.ts
@@ -1,0 +1,12 @@
+import type {PageContributions} from '../../../../index.d';
+
+
+export function appendCssToHeadEnd(
+	url :string,
+	pageContributions :PageContributions
+) {
+  pageContributions.headEnd = [
+    ...(pageContributions.headEnd || []),
+    `<link href="${url}" rel="stylesheet" type="text/css" />\n`
+  ];
+};
